requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
ccxt==4.1.13
pandas==2.1.1
numpy==1.24.3
websockets==11.0.3
python-binance==1.0.19
APScheduler==3.10.4
Flask-Cors==4.0.0
gunicorn==21.2.0 